added some utils

+++ philo/srcs/utils.c
@@ -6,7 +6,7 @@
 /*   By: ting <ting@student.42singapore.sg>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/18 15:35:56 by ting              #+#    #+#             */
-/*   Updated: 2024/04/20 00:00:16 by ting             ###   ########.fr       */
+/*   Updated: 2024/04/20 15:26:17 by ting             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void	ft_error(char *str)
 	exit(1);
 }
 
-u_int64_t   get_time(void)
+long   get_time(void)
 {
     struct timeval   time;
 
@@ -26,5 +26,5 @@ u_int64_t   get_time(void)
     {
         ft_error("gettimeofday Error");
     }
-    return((time.tv_sec * (u_int64_t)1000) + (time.tv_usec / 1000));
+    return((time.tv_sec * 1000) + (time.tv_usec / 1000));
 }
